index page cleanup 3
