+++ src/lib/shared/components/featured-content/FeaturedContent.svelte
@@ -19,7 +19,7 @@
   setInterval(advanceIndex, 2000);
 </script>
 
-<div class="w-[75%]">
+<div class="w-[100%]">
   <h2
     id="featured"
     class="font-bold text-2xl md:text-4xl tracking-tight mb-4 max-w-5xl text-black dark:text-white"
